Move production types to the `EssentialFeed` production module
