+++ EssentialFeed/EssentialFeed/Feed Presentation/FeedPresenter.swift
@@ -1,0 +1,81 @@
+//
+//  FeedPresenter.swift
+//  EssentialFeed
+//
+//  Created by Andrey on 9/8/22.
+//
+
+import Foundation
+
+public struct FeedViewModel {
+    public let feed: [FeedImage]
+}
+
+public protocol FeedView {
+    func display(_ viewModel: FeedViewModel)
+}
+
+public struct FeedErrorViewModel {
+    public let message: String?
+    
+    static var noError: FeedErrorViewModel {
+        FeedErrorViewModel(message: .none)
+    }
+    
+    static func error(message: String) -> FeedErrorViewModel {
+        FeedErrorViewModel(message: message)
+    }
+}
+
+public protocol FeedErrorView {
+    func display(_ viewModel: FeedErrorViewModel)
+}
+
+public struct FeedLoadingViewModel {
+    public let isLoading: Bool
+}
+
+public protocol FeedLoadingView {
+    func display(_ viewModel: FeedLoadingViewModel)
+}
+
+public final class FeedPresenter {
+    private let feedView: FeedView
+    private let loadingView: FeedLoadingView
+    private let errorView: FeedErrorView
+    
+    public init(feedView: FeedView, loadingView: FeedLoadingView, errorView: FeedErrorView) {
+        self.feedView = feedView
+        self.loadingView = loadingView
+        self.errorView = errorView
+    }
+    
+    public static var title: String {
+        NSLocalizedString("FEED_VIEW_TITLE",
+            tableName: "Feed",
+            bundle: Bundle(for: FeedPresenter.self),
+            comment: "")
+    }
+    
+    private var errorMessage: String {
+        NSLocalizedString("FEED_VIEW_CONNECTION_ERROR",
+            tableName: "Feed",
+            bundle: Bundle(for: FeedPresenter.self),
+            comment: "")
+    }
+    
+    public func didStartLoadingFeed() {
+        errorView.display(.noError)
+        loadingView.display(.init(isLoading: true))
+    }
+    
+    public func didFinishLoadingFeed(with feed: [FeedImage]) {
+        feedView.display(FeedViewModel(feed: feed))
+        loadingView.display(FeedLoadingViewModel(isLoading: false))
+    }
+    
+    public func didFinishLoadingFeed(with error: Error) {
+        errorView.display(.error(message: errorMessage))
+        loadingView.display(FeedLoadingViewModel(isLoading: false))
+    }
+}
